Adds path support for SNAPSHOT_APP_IDENTIFIER
If the SNAPSHOT_APP_IDENTIFIER environment variable is set, the runner
will search for the .app with the given identifer, rather than taking
the first .app it can find.

If the environment variable isn't set, the previous behavior of just
grabbing the first non-watchkit .app still occurs.

This is relevant for schemes which generate more than one application
bundle, such as a test rig for running unit tests.